Corrected the code based on review feedbacks. Separated the composable part from the date logic part. Also some minor changes for good practices and the view model documentation.
